chore(organization): Reorganized store devtools code for better interop with SystemJS loader and npmcdn
