+++ src/index.ts
@@ -1,3 +1,8 @@
-export * from './store';
-export {LogMonitor} from './monitors/log-monitor';
-export * from './monitors/dock-monitor';
+export {instrumentStore} from './store/instrument';
+export {LiftedState} from './store/reducer';
+export {StoreDevtools} from './store/devtools';
+
+export {DockMonitor} from './monitors/dock-monitor/dock-monitor';
+export {dockReducer} from './monitors/dock-monitor/reducer';
+
+export {LogMonitor} from './monitors/log-monitor/log-monitor';
